\name{mcll-package}
\alias{mcll-package}

\docType{package}

\title{
Monte Carlo local likelihood estimation
}

\description{
This package provides estimates of parameters and standard errors using a Monte Carlo local likelihood method.
The pacakge \code{mcll} is based on Jeon, Kaufman, and Rabe-Hesketh (2014). 

}

\details{
\tabular{ll}{
Package: \tab mcll\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2014-3-1\cr
License: \tab \acronym{GPL} \cr
}
}

\author{ 
Minjeong Jeon <jeon.117@osu.edu>, Cari Kaufman <cgk@berkeley.edu>, and Sophia Rabe-Hesketh <sophiarh@berkeley.edu> 

Maintainer:   Minjeong Jeon <jeon.117@osu.edu>
}

\references{

  Jeon, M., Kaufman, C., and Rabe-Hesketh, S. (2014). 
  Monte Carlo local likelihood for approximate MLE of complex models. Under revision. 

}

\keyword{package}

\keyword{mcll}
