\name{samp}

\alias{samp}

\title{Posterior samples of model parameters for the salamander mating model}

\description{
These are 3,000 posterior samples of the six model parameters (four fixed effects and two log standard deviation parameters).
For the model and priors, see Section 4 in Jeon et al. (2012).  
}

\format{
 A data matrix with 3,000 posterior samples for the six parameters.
}

\references{

  Jeon, M., Kaufman, C., and Rabe-Hesketh, S. (2012). Monte Carlo local likelihood 
  for estimating generalized linear mixed models. Submitted for publication. 

}

\examples{

data(samp)
str(samp)


}


\keyword{datasets} 
