\name{mcll-package}
\alias{mcll-package}

\docType{package}

\title{
Monte Carlo local likelihood estimation
}

\description{
This package provides estimates of parameters and standard errors using a Monte Carlo local likelihood method.
The pacakge \code{mcll} is based on Jeon, Kaufman, and Rabe-Hesketh (2012). 

}

\details{
\tabular{ll}{
Package: \tab mcll\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-28\cr
License: \tab \acronym{GPL} \cr
}
}

\author{ 
Minjeong Jeon <mjj@berkeley.edu>, Cari Kaufman <cgk@berkeley.edu>, and Sophia Rabe-Hesketh <sophiarh@berkeley.edu> 

Maintainer:   Minjeong Jeon <mjj@berkeley.edu>
}

\references{

  Jeon, M., Kaufman, C., and Rabe-Hesketh, S. (2012). 
  Monte Carlo local likelihood for estimating generalized linear mixed models. Submitted for pulication. 

}

\keyword{package}

\keyword{mcll}

